removes signup, welcome, and private router
