--- src/containers/SigninPage/Login.jsx
@@ -1,59 +1,0 @@
-import React from 'react';
-import { useRef, useState } from 'react';
-import { Form, Button, Card, Alert } from 'react-bootstrap';
-import 'bootstrap/dist/css/bootstrap.min.css';
-import { useAuth } from '../../context/AuthContext';
-import { Link, useHistory } from 'react-router-dom';
-
-function Login() {
-  const emailRef = useRef();
-  const passwordRef = useRef();
-  const { login } = useAuth();
-  const [error, setError] = useState('');
-  const [loading, setLoading] = useState(false);
-  const history = useHistory();
-
-  async function handleSubmit(event) {
-    event.preventDefault();
-
-    try {
-      setError('');
-      setLoading(true);
-      await login(emailRef.current.value, passwordRef.current.value);
-      history.push('/welcome');
-    } catch {
-      setError('Failed to sign in');
-    }
-
-    setLoading(false);
-  }
-
-  return (
-    <>
-      <Card>
-        <Card.Body>
-          <h2 className='text-center mb-4'>Log In</h2>
-          {error && <Alert variant='danger'>{error}</Alert>}
-          <Form onSubmit={handleSubmit}>
-            <Form.Group id='email'>
-              <Form.Label>Email</Form.Label>
-              <Form.Control type='email' ref={emailRef} required />
-            </Form.Group>
-            <Form.Group id='password'>
-              <Form.Label>Password</Form.Label>
-              <Form.Control type='password' ref={passwordRef} required />
-            </Form.Group>
-            <Button type='submit' className='w-100' disabled={loading}>
-              Log In
-            </Button>
-          </Form>
-        </Card.Body>
-      </Card>
-      <div className='w-100 text-center mt-2'>
-        Don't have an account? <Link to='/signup'>Sign up</Link>
-      </div>
-    </>
-  );
-}
-
-export default Login;
