--- src/containers/SigninPage/PrivateRoute.jsx
@@ -1,18 +1,0 @@
-import React from 'react';
-import { Route, Redirect } from 'react-router-dom';
-import { useAuth } from '../../context/AuthContext';
-
-function PrivateRoute({ component: Component, ...rest }) {
-  const { currentUser } = useAuth();
-
-  return (
-    <Route
-      {...rest}
-      render={(props) => {
-        return currentUser ? <Component {...props} /> : <Redirect to='/' />;
-      }}
-    ></Route>
-  );
-}
-
-export default PrivateRoute;
