--- src/containers/SigninPage/SignUp.jsx
@@ -1,72 +1,0 @@
-import React from 'react';
-import { useRef, useState } from 'react';
-import { Form, Button, Card, Alert } from 'react-bootstrap';
-import 'bootstrap/dist/css/bootstrap.min.css';
-import { useAuth } from '../../context/AuthContext';
-import { Link, useHistory } from 'react-router-dom';
-
-function SignUp() {
-  const emailRef = useRef();
-  const passwordRef = useRef();
-  const passwordConfirmRef = useRef();
-  const { signup } = useAuth();
-  const [error, setError] = useState('');
-  const [loading, setLoading] = useState(false);
-  const history = useHistory();
-
-  async function handleSubmit(event) {
-    event.preventDefault();
-
-    if (passwordRef.current.value !== passwordConfirmRef.current.value) {
-      return setError('Passwords do not match');
-    }
-
-    try {
-      setError('');
-      setLoading(true);
-      await signup(
-        emailRef.current.value,
-        passwordRef.current.value,
-        passwordConfirmRef.current.value
-      );
-      history.push('/welcome');
-    } catch {
-      setError('Failed to create an account');
-    }
-
-    setLoading(false);
-  }
-
-  return (
-    <>
-      <Card>
-        <Card.Body>
-          <h2 className='text-center mb-4'>Sign Up</h2>
-          {error && <Alert variant='danger'>{error}</Alert>}
-          <Form onSubmit={handleSubmit}>
-            <Form.Group id='email'>
-              <Form.Label>Email</Form.Label>
-              <Form.Control type='email' ref={emailRef} required />
-            </Form.Group>
-            <Form.Group id='password'>
-              <Form.Label>Password</Form.Label>
-              <Form.Control type='password' ref={passwordRef} required />
-            </Form.Group>
-            <Form.Group id='password-confirm'>
-              <Form.Label>Password Confirmation</Form.Label>
-              <Form.Control type='password' ref={passwordConfirmRef} required />
-            </Form.Group>
-            <Button type='submit' className='w-100' disabled={loading}>
-              Sign Up
-            </Button>
-          </Form>
-        </Card.Body>
-      </Card>
-      <div className='w-100 text-center mt-2'>
-        Already have an account? <Link to='/'>Log in</Link>
-      </div>
-    </>
-  );
-}
-
-export default SignUp;
