+++ src/containers/SigninPage/SigninPage.jsx
@@ -1,32 +1,56 @@
-import './SigninPage.css';
-import { Container } from 'react-bootstrap';
+import React from 'react';
+import { useRef, useState } from 'react';
+import { Form, Button, Card, Alert } from 'react-bootstrap';
 import 'bootstrap/dist/css/bootstrap.min.css';
-import SignUp from './SignUp';
-import Login from './Login';
-import Welcome from './Welcome';
-import { AuthProvider } from '../../context/AuthContext';
-import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
-import PrivateRoute from './PrivateRoute';
+import { useAuth } from '../../context/AuthContext';
+import { useHistory } from 'react-router-dom';
 
-const SigninPage = () => {
+function SigninPage() {
+  const emailRef = useRef();
+  const passwordRef = useRef();
+  const { login } = useAuth();
+  const [error, setError] = useState('');
+  const [loading, setLoading] = useState(false);
+  const history = useHistory();
+
+  async function handleSubmit(event) {
+    event.preventDefault();
+
+    try {
+      setError('');
+      setLoading(true);
+      await login(emailRef.current.value, passwordRef.current.value);
+      history.push('/admin');
+    } catch {
+      setError('Failed to sign in');
+    }
+
+    setLoading(false);
+  }
+
   return (
-    <Container
-      className='d-flex align-items-center justify-content-center'
-      style={{ minHeight: '100vh' }}
-    >
-      <div className='w-100' style={{ maxWidth: '400px' }}>
-        <Router>
-          <AuthProvider>
-            <Switch>
-              <Route exact path='/' component={Login} />
-              <Route path='/signup' component={SignUp} />
-              <PrivateRoute path='/welcome' component={Welcome} />
-            </Switch>
-          </AuthProvider>
-        </Router>
-      </div>
-    </Container>
+    <div className='w-100' style={{ maxWidth: '400px' }}>
+      <Card>
+        <Card.Body>
+          <h2 className='text-center mb-4'>Log In</h2>
+          {error && <Alert variant='danger'>{error}</Alert>}
+          <Form onSubmit={handleSubmit}>
+            <Form.Group id='email'>
+              <Form.Label>Email</Form.Label>
+              <Form.Control type='email' ref={emailRef} required />
+            </Form.Group>
+            <Form.Group id='password'>
+              <Form.Label>Password</Form.Label>
+              <Form.Control type='password' ref={passwordRef} required />
+            </Form.Group>
+            <Button type='submit' className='w-100' disabled={loading}>
+              Log In
+            </Button>
+          </Form>
+        </Card.Body>
+      </Card>
+    </div>
   );
-};
+}
 
 export default SigninPage;
